Names of the Distinct Second-Derivative Components

The keys of struct_d2matrix and struct_d2logdet: one per unordered pair of free values, diagonal first and then the off-diagonal pairs in lexicographic order.

Usage

struct_pair_names(s)

Arguments

s

An object inheriting from class covstruct.

Value

A character vector of length \(d(d+1)/2\).

Details

This exists so that nothing has to recover a pair by splitting a key apart. Generating the keys and the index pairs from one enumeration cannot be fooled by a free name that contains the separator, which splitting can. Use struct_pair_indices for the pairs themselves.

See also

Examples

struct_pair_names(log_cholesky(2))
#> [1] "log_L1:log_L1" "log_L2:log_L2" "L2.1:L2.1"     "log_L1:log_L2"
#> [5] "log_L1:L2.1"   "log_L2:L2.1"